ref: update linkableapi

+++ src/com/troblecodings/linkableapi/Linkingtool.java
@@ -24,8 +24,11 @@ import net.minecraft.world.World;
 
 public class Linkingtool extends Item {
 
-    private final BiPredicate<World, BlockPos> predicate;
-    private final Predicate<TileEntity> predicateSet;
+    protected static final String LINKINGTOOL_TAG = "linkingToolTag";
+
+    protected final BiPredicate<World, BlockPos> predicate;
+    protected final Predicate<TileEntity> predicateSet;
+    protected final TaggableFunction tagFromFunction;
 
     public Linkingtool(final ItemGroup tab, final BiPredicate<World, BlockPos> predicate) {
         this(tab, predicate, _u -> true);
@@ -33,9 +36,16 @@ public class Linkingtool extends Item {
 
     public Linkingtool(final ItemGroup tab, final BiPredicate<World, BlockPos> predicate,
             final Predicate<TileEntity> predicateSet) {
-        super(new Properties().group(tab));
+        this(tab, predicate, predicateSet, (_u1, _u2, _u3) -> {
+        });
+    }
+
+    public Linkingtool(final ItemGroup tab, final BiPredicate<World, BlockPos> predicate,
+            final Predicate<TileEntity> predicateSet, final TaggableFunction function) {
+        super(new Properties().group(tab).setNoRepair().defaultMaxDamage(64).maxStackSize(1));
         this.predicate = predicate;
         this.predicateSet = predicateSet;
+        this.tagFromFunction = function;
     }
 
     @Override
@@ -43,49 +53,76 @@ public class Linkingtool extends Item {
         final World levelIn = ctx.getWorld();
         final EntityPlayer player = ctx.getPlayer();
         final BlockPos pos = ctx.getPos();
+        if (player == null)
+            return EnumActionResult.FAIL;
         if (levelIn.isRemote)
             return EnumActionResult.PASS;
         final TileEntity entity = levelIn.getTileEntity(pos);
+        final NBTTagCompound itemTag = getOrCreateForStack(stack);
+        final NBTTagCompound toolTag = itemTag.getCompound(LINKINGTOOL_TAG);
         if (entity instanceof ILinkableTile && this.predicateSet.apply(entity)) {
             final ILinkableTile controller = (ILinkableTile) entity;
             if (!player.isSneaking()) {
                 final NBTTagCompound comp = stack.getTag();
-                if (comp == null) {
+                if (toolTag == null) {
                     message(player, "lt.notset", pos.toString());
                     return EnumActionResult.PASS;
                 }
                 final BlockPos lpos = NBTUtil.readBlockPos(comp);
-                if (controller.link(lpos)) {
+                if (controller.link(lpos, toolTag)) {
                     message(player, "lt.linkedpos", pos.getX(), pos.getY(), pos.getZ());
-                    stack.setTag(null);
+                    removeToolTag(stack);
                     message(player, "lt.reset");
+                    stack.damageItem(1, player);
                     return EnumActionResult.FAIL;
                 }
                 message(player, "lt.notlinked");
                 message(player, "lt.notlinked.msg");
                 return EnumActionResult.FAIL;
             } else {
+                if (controller.canBeLinked() && predicate.test(levelIn, pos)) {
+                    final boolean containsPos =
+                            toolTag.hasKey("X") && toolTag.hasKey("Y") && toolTag.hasKey("Z");
+                    if (containsPos) {
+                        message(player, "lt.setpos.msg");
+                        return EnumActionResult.FAIL;
+                    }
+                    final NBTTagCompound newToolTag = NBTUtil.writeBlockPos(pos);
+                    tagFromFunction.test(levelIn, pos, newToolTag);
+                    itemTag.setTag(LINKINGTOOL_TAG, newToolTag);
+                    message(player, "lt.setpos", pos.getX(), pos.getY(), pos.getZ());
+                    message(player, "lt.setpos.msg");
+                    return EnumActionResult.SUCCESS;
+                }
                 if (controller.hasLink() && controller.unlink()) {
                     message(player, "lt.unlink");
+                    return EnumActionResult.SUCCESS;
                 }
             }
             return EnumActionResult.SUCCESS;
         } else if (predicate.test(levelIn, pos)) {
-            if (stack.getTag() != null) {
+            final boolean containsPos =
+                    toolTag.hasKey("X") && toolTag.hasKey("Y") && toolTag.hasKey("Z");
+            if (containsPos) {
                 message(player, "lt.setpos.msg");
                 return EnumActionResult.FAIL;
             }
-            final NBTTagCompound comp = NBTUtil.writeBlockPos(pos);
-            stack.setTag(comp);
+            final NBTTagCompound newToolTag = NBTUtil.writeBlockPos(pos);
+            tagFromFunction.test(levelIn, pos, newToolTag);
+            itemTag.setTag(LINKINGTOOL_TAG, newToolTag);
             message(player, "lt.setpos", pos.getX(), pos.getY(), pos.getZ());
             message(player, "lt.setpos.msg");
             return EnumActionResult.SUCCESS;
         } else if (player.isSneaking() && stack.getTag() != null) {
-            stack.setTag(null);
+            removeToolTag(stack);
             message(player, "lt.reset");
             return EnumActionResult.SUCCESS;
         }
         return EnumActionResult.FAIL;
+    }
+
+    public void removeToolTag(final ItemStack stack) {
+        getOrCreateForStack(stack).removeTag(LINKINGTOOL_TAG);
     }
 
     @Override
@@ -103,10 +140,16 @@ public class Linkingtool extends Item {
         tooltip(tooltip, "lt.notlinked.msg");
     }
 
-    @SuppressWarnings({
-            "rawtypes", "unchecked"
-    })
-    public void tooltip(final List list, final String text, final Object... obj) {
+    public static NBTTagCompound getOrCreateForStack(final ItemStack stack) {
+        NBTTagCompound tag = stack.getTag();
+        if (tag == null) {
+            tag = new NBTTagCompound();
+            stack.setTag(tag);
+        }
+        return tag;
+    }
+
+    public void tooltip(final List<ITextComponent> list, final String text, final Object... obj) {
         list.add(getComponent(text, obj));
     }
